+++ src/routes/socketEvents.ts
@@ -3,16 +3,12 @@
 import { Role, type User } from '@prisma/client';
 import { Server } from 'socket.io';
 import Logger from '../utils/logger';
-import {
-    ClientToServerEvents,
-    IoEvent,
-    IoClientEvent,
-    ServerToClientEvents,
-    Action
-} from './socketEventTypes';
+import { ClientToServerEvents, IoEvent, IoClientEvent, ServerToClientEvents } from './socketEventTypes';
 import StudentGroup from '../models/StudentGroup';
-import prisma from '../prisma';
 import { hasElevatedAccess } from '../models/User';
+import onAction from './event-handlers/action.handler';
+import onJoinRoom from './event-handlers/joinRoom.handler';
+import onLeaveRoom from './event-handlers/leaveRoom.handler';
 
 export enum IoRoom {
     ADMIN = 'admin',
@@ -35,156 +31,11 @@ const EventRouter = (io: Server<ClientToServerEvents, ServerToClientEvents>) => 
             io.to(IoRoom.ADMIN).emit(IoEvent.CONNECTED_CLIENTS, { rooms: rooms, type: 'full' });
         }
         if (hasElevatedAccess(user.role)) {
-            socket.on(IoClientEvent.ACTION, (navRequest, onDone) => {
-                if (user.role === Role.ADMIN) {
-                    socket
-                        .to([...navRequest.roomIds, ...navRequest.userIds])
-                        .except(socket.id)
-                        .emit(IoEvent.ACTION, navRequest.action);
-                    return onDone(true);
-                }
-                // check access first
-                (navRequest.roomIds.length > 0
-                    ? prisma.studentGroup
-                          .findMany({
-                              where: {
-                                  id: { in: navRequest.roomIds },
-                                  users: {
-                                      some: {
-                                          userId: user.id,
-                                          isAdmin: true
-                                      }
-                                  }
-                              },
-                              select: {
-                                  id: true
-                              }
-                          })
-                          .then((sg) => {
-                              return sg.map((group) => group.id);
-                          })
-                    : Promise.resolve([])
-                ).then((groupIds) => {
-                    return (
-                        navRequest.userIds.length > 0
-                            ? prisma.studentGroup
-                                  .findMany({
-                                      where: {
-                                          AND: [
-                                              {
-                                                  users: {
-                                                      some: {
-                                                          userId: { in: navRequest.userIds }
-                                                      }
-                                                  }
-                                              },
-                                              {
-                                                  users: {
-                                                      some: {
-                                                          userId: user.id,
-                                                          isAdmin: true
-                                                      }
-                                                  }
-                                              }
-                                          ]
-                                      },
-                                      select: {
-                                          users: {
-                                              select: {
-                                                  userId: true
-                                              }
-                                          }
-                                      }
-                                  })
-                                  .then((studentGroups) => {
-                                      return studentGroups
-                                          .flatMap((group) => group.users.flatMap((user) => user.userId))
-                                          .filter((id) => navRequest.userIds.includes(id));
-                                  })
-                            : Promise.resolve([])
-                    ).then((userIds) => {
-                        const audience = [...userIds, ...groupIds];
-                        if (audience.length > 0) {
-                            socket.to(audience).except(socket.id).emit(IoEvent.ACTION, navRequest.action);
-                        }
-                        onDone(true);
-                    });
-                });
-            });
+            socket.on(IoClientEvent.ACTION, onAction(user, socket));
         }
         socket.join(IoRoom.ALL);
-        socket.on(IoClientEvent.JOIN_ROOM, (roomId: string, callback: (joined: boolean) => void) => {
-            if (user.role === Role.ADMIN) {
-                socket.join(roomId);
-                return callback(true);
-            }
-            StudentGroup.findModel(user, roomId).then((group) => {
-                if (group) {
-                    socket.join(roomId);
-                    callback(true);
-                } else {
-                    if (user.role === Role.TEACHER) {
-                        prisma.studentGroup
-                            .findFirst({
-                                where: {
-                                    users: {
-                                        some: {
-                                            AND: [
-                                                {
-                                                    userId: user.id,
-                                                    isAdmin: true
-                                                },
-                                                {
-                                                    userId: roomId
-                                                }
-                                            ]
-                                        }
-                                    }
-                                }
-                            })
-                            .then((userRoom) => {
-                                if (userRoom) {
-                                    socket.join(roomId);
-                                    callback(true);
-                                } else {
-                                    callback(false);
-                                }
-                            });
-                    }
-                }
-            });
-        });
-        socket.on(IoClientEvent.LEAVE_ROOM, (roomId: string, callback: (left: boolean) => void) => {
-            if (hasElevatedAccess(user.role)) {
-                socket.leave(roomId);
-                return callback(true);
-            }
-            // students can only leave rooms where they are not part of.
-            prisma.userStudentGroup
-                .findMany({
-                    where: {
-                        studentGroupId: roomId
-                    },
-                    select: {
-                        userId: true
-                    },
-                    distinct: ['userId']
-                })
-                .then((roomMembers) => {
-                    // leave the room if:
-                    // - the roomId is associated to a room, when it has at least one member
-                    // - the user is not part of the room
-                    if (
-                        roomMembers.length > 0 &&
-                        !roomMembers.map((member) => member.userId).includes(user.id)
-                    ) {
-                        socket.leave(roomId);
-                        callback(true);
-                    } else {
-                        callback(false);
-                    }
-                });
-        });
+        socket.on(IoClientEvent.JOIN_ROOM, onJoinRoom(user, socket));
+        socket.on(IoClientEvent.LEAVE_ROOM, onLeaveRoom(user, socket));
         const groups = await StudentGroup.all(user);
         const groupIds = groups.map((group) => group.id);
         if (groupIds.length > 0) {
